give up all hope of a post stdlib 4.1.0 release happening
This module is dependant the behavior of the stdlib merge() function as
of commit f496005bf3db8a5202bf9c16daf9a524b178c67a, which was merged
after lastest (4.1.0) release.  Since that release is almost a year old
now, it's looking unlikely that there will be another release in the
near future.

As a workaround, the merge() function has been copied into this module as
pureftpd_merge() and the Module dep as been reduced to stdlib >= 4.0.0.