remove bundled pureftpd_merge() function
The pureftpd_merge() function included in this module is a copy of the
stdlib merge() function from commit id
f496005bf3db8a5202bf9c16daf9a524b178c67a.  This version of merge()
includes a critical bug fix that was commit after the stdlib 4.1.0
release.  Now that stdlib 4.2.0 has been released (finally!) with the
needed fix, bundling a version of the merge() function in this module is
no longer necessary.